Summer Solstice 2021 update
For the Summer Solstice 2021, current public release is in sync with Patreon Release. Enjoy!
This is A major update with lots of under the hood changes, focused on efficiency and simpler structure for the system (check the changelog for more details).
Next round of developments will focus on new features and geometrical series, to be released on Patreon first, on the Expert version tiers.